Face detection/requirement.txt
opencv-python
streamlit
pillow
numpy
tensorflow